+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - Added InOutSingleChannel
 - Add optional `config_settings` property to quam components indicating that they should be called before/after other components when generating QUA configuration
+- Added `InOutIQChannel.measure_accumulated/sliced`
 
 ### Changed
 - Changed `InOutIQChannel.input_offset_I/Q` to `InOutIQChannel.opx_input_offset_I/Q`
@@ -17,6 +18,7 @@
   This happens if channel.id is not set, and channel.parent does not have a name either
 - `Pulse.axis_angle` is now in radians instead of degrees.
 - Channel offsets (e.g. `SingleChannel.opx_output_offset`) is None by default (see note in Fixed)
+- Replaced `InOutIQChannel.measure` kwargs `I_var` and `Q_var` by `qua_vars` tuple
 
 ### Fixed
 - Don't raise instantiation error when required_type is not a class